+++ src/test/java/racingcar/domain/cars/CarNamesTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.cars;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static racingcar.config.RacingCarException.CAR_NAME_DUPLICATED;
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import racingcar.domain.cars.CarNames;
 
 class CarNamesTest {
