+++ src/test/java/racingcar/domain/movement/MovementStrategyTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.movement;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import racingcar.TestRacingCarConfig;
+import racingcar.domain.movement.MovementStrategy;
 
 class MovementStrategyTest {
